Add files via upload
A very Basic coding for testing. After getting the dataa from dht we send it to firebase rtdb and after that on react we get them for monitoring. The complicated version is made on PlantAPP -you can find it on my repository but it's made for android devices not React-. To make project professional 
1. you can add more sensors.
2. you can make a control system to turn on/off the sensors.
3. you can add live stream system via esp cam/using firebase stroage as different not rtdb.
4. you can chage css for make it look good.